Move kv cache related to resource manager
